Refactor to dataclasses and avoid extensive open/close operations
Avoid the repeated use of open/close operations due to the use of
context manager on every read operation. Use the class as a whole as
context manager instead, to ensure close on exit.

Refactor to using data classes to improve code structure and
readability, and better compatibility with database operations.